image: added multiple ubi configuration capability
The Ventana product family supports multiple flash chips that fall into
two flash geometry configurations which we simply call 'normal' and 'large'.

The default is the normal layout and you will need to specify UBI_CONFIG=large
to override this in your MACHINE or global configuration if you want to build
the other.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>