test_video: add detection of PAL source
The test_video script is a front-end for creating and launching
gstreamer pipelines for video loopback.  When specifying an input
source, It is important to not launch the pipeline until the source has
detected an input otherwise the pipeline will not use the correct modes
and formats.  This detection uses v4l2-ctl or sysfs depending on the
input type (analog vs digital).  For analog, only NTSC was being accepted
as a valid input.  This patch adds the other possibility, PAL to that check.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>