Updated: compat-wireless recipe
This updates compat-wireless to latest stable compat-wireless
release. This includes refreshing the defconfig to include iwlwifi
drivers. It should be noted, however, that the iwlwifi driver requires
MSI to be turned on to be usable. However, ath10k requries that we have
MSI disabled. We have opted to disable MSI until we find a resolution
that allows both ath10k and iwlwifi to work in conjunction with one
another.

Signed-off-by: Pushpal Sidhu <psidhu@gateworks.com>