recipes-kernel/compat-wireless: use internal regdb
Instead of using the crda package which is often times outdated (it was
last updated in 20130118), we have opted to use the internal regdb. This
means that the regulatory domain is built into the kernel itself.

The current regdb in use allows for the 80MHz channels to be
utilized. For exact details on which channels we have configured in our
database, see 79d11dbf of wireless-regdb.git.

Since this update is moving from crda to  the internal regdb, we have
also removed the crda package in this commit.

Signed-off-by: Pushpal Sidhu <psidhu@gateworks.com>