init-ifupdown: change wpa-driver from wext to nl80211
wext has been deprecated for some time as stated in the wpa_supplicant
defconfig:

Note: WEXT is deprecated in the current Linux kernel version and no new
functionality is added to it. nl80211-based interface is the new
replacement for WEXT and its use allows wpa_supplicant to properly control
the driver to improve existing functionality like roaming and to support new
functionality.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>